Move serverId utils to internal package
