+++ core/trace/internal/util/server_id.go
@@ -1,17 +1,18 @@
-package trace
+package util
 
 import "context"
 
 type serverIdKeyType int
+
 const serverIdKey serverIdKeyType = iota
 
 // This will be called from the Dynatrace propagator
 //nolint:deadcode,unused
-func setServerIdOnContext(ctx context.Context, serverId int32) context.Context {
+func SetServerIdOnContext(ctx context.Context, serverId int32) context.Context {
 	return context.WithValue(ctx, serverIdKey, serverId)
 }
 
-func getServerIdFromContext(ctx context.Context) int32 {
+func GetServerIdFromContext(ctx context.Context) int32 {
 	if serverId, ok := ctx.Value(serverIdKey).(int32); ok {
 		return serverId
 	}
